backend/requirements.txt
fastapi
uvicorn
sqlalchemy
networkx
requests
gitpython
python-multipart
